fix bug when processing input file comments (#2808)
and move 'parse_inputfile()' to util.py
